+++ clx_task_management/models/project.py
@@ -1,7 +1,14 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo, CLx Media
 # See LICENSE file for full copyright & licensing details.
-from odoo import fields, models
+from odoo import fields, models, _
+from odoo.exceptions import UserError
+
+
+class projectTaskType(models.Model):
+    _inherit = 'project.task.type'
+
+    demo_data = fields.Boolean()
 
 
 class ProjectProject(models.Model):
@@ -20,6 +27,13 @@ class ProjectTask(models.Model):
     req_type = fields.Selection([('new', 'New'), ('update', 'Update')],
                                 string='Request Type')
     sub_task_id = fields.Many2one('sub.task', string="Sub Task from Master Table")
+    team_id = fields.Many2one('clx.team', string='Team')
+    team_members_ids = fields.Many2many('res.users', string="Team Members")
+
+    def unlink(self):
+        for task in self:
+            raise UserError(_('You can Not Delete the Task {} please contact Administrator.').format(task.name))
+        return super(ProjectTask, self).unlink()
 
     def create_sub_task(self, task, project_id):
         stage_id = self.env.ref('clx_task_management.clx_project_stage_1')
@@ -37,7 +51,6 @@ class ProjectTask(models.Model):
     def write(self, vals):
         res = super(ProjectTask, self).write(vals)
         sub_task_obj = self.env['sub.task']
-        project_task = self.env['project.task']
         stage_id = self.env['project.task.type'].browse(vals.get('stage_id'))
         complete_stage = self.env.ref('clx_task_management.clx_project_stage_8')
         if vals.get('stage_id', False) and stage_id.id == complete_stage.id:
@@ -45,5 +58,5 @@ class ProjectTask(models.Model):
                 dependency_tasks = sub_task_obj.search([('dependency_ids', 'in', self.sub_task_id.ids)])
                 for task in dependency_tasks:
                     vals = self.create_sub_task(task, self.project_id)
-                    project_task.create(vals)
+                    self.create(vals)
         return res